requirements.txt
numpy
insightface
onnxruntime
opencv-python-headless
google-cloud-storage